requirements.txt
Flask==2.2.2
Pillow==9.2.0
openpyxl==3.0.10